Extract adaptive signal filters
